Sessions - view all changes action uses the diff editor + navigation tree
